migrate NewList to prebuilts
